% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.cutByEvent.R
\name{cross.cutByEvent}
\alias{cross.cutByEvent}
\title{Cut input by event marks, return 'channels' as result of concatenating}
\usage{
cross.cutByEvent(data, events, shiftT = 0, shiftF = 0)
}
\description{
Cut input by event marks, return 'channels' as result of concatenating
}
