% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.epochByEvent.R
\name{cross.epochByEvent}
\alias{cross.epochByEvent}
\title{Filters data stream cutting marked blocks}
\usage{
cross.epochByEvent(data, events, shiftT = 0, shiftF = 0)
}
\arguments{
\item{data}{data stream}

\item{events}{event stream}
}
\description{
Filters data stream cutting marked blocks
}
