% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronization.R
\name{cross.syncByTimestamp}
\alias{cross.syncByTimestamp}
\title{Perform synchronization of two streams using RTC timestamps}
\usage{
cross.syncByTimestamp(inputA, inputB, delayBuffer)
}
\arguments{
\item{inputA}{First input}

\item{inputB}{Second input}
}
\value{
List with pipes
}
\description{
Simple synchronization, recommended for synchronization streams generated (not recorded) on same machine.
In other cases can lead to constant desynchronization.
}
