% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drain.channelsRecorder}
\alias{drain.channelsRecorder}
\title{Terminating pipe that records all signal to one big buffer.}
\usage{
drain.channelsRecorder(input)
}
\arguments{
\item{input}{Input stream.}
}
\value{
function that returns recorded signal
}
\description{
Terminating pipe that records all signal to one big buffer.
}

