% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pipe.bandFilter}
\alias{pipe.bandFilter}
\title{Simple filter pipe that applies high-pass, low-pass and notch filters}
\usage{
pipe.bandFilter(input, lowFreq, highFreq, notch, order = 2)
}
\arguments{
\item{input}{Pipe connected to}

\item{lowFreq}{Frequency of high-pass filter}

\item{highFreq}{Frequency of low-pass filter}

\item{notch}{Frequency for notch filter}
}
\value{
Constructed pipe
}
\description{
Simple filter pipe that applies high-pass, low-pass and notch filters
}

