% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.eventsEquals.R
\name{pipe.eventsEquals}
\alias{pipe.eventsEquals}
\title{Transforms events to boolean
By rule event == match}
\usage{
pipe.eventsEquals(input, match)
}
\arguments{
\item{input}{event stream}

\item{match}{value to match}
}
\value{
event stream
}
\description{
Transforms events to boolean
By rule event == match
}
