% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.spectrogram.R
\name{pipe.spectrogram}
\alias{pipe.spectrogram}
\title{Extracts spectrograms for windows}
\usage{
pipe.spectrogram(input)
}
\arguments{
\item{input}{Input pipe}

\item{choose}{DataFrame with columns 'channel' and 'frequency'}
}
\value{
Row where each element corresponds to row from choose and equals to power of channel
}
\description{
Extracts spectrograms for windows
}
