% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pipeline}
\alias{pipeline}
\title{Combine pipes to chain, returns it's output}
\usage{
pipeline(..., .env = parent.frame())
}
\arguments{
\item{...}{List of pipes, I recommend to skip first argument for pipe}

\item{.env}{Environment for evaluating pipes, for experts only}
}
\value{
Output of last pipe
}
\description{
Allows nice descriptive specification of pipeline.
Iterates over arguments and constructs pipeline where output from \emph{n-1} filter directed to input of \emph{n} filter.
}
\examples{
pipeline(
 pipe.applyFilter(input, f1),
 pipe.applyFilter(, f2)
)
}

