% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{processor}
\alias{processor}
\title{Wraps all necessary handlers to pipeline function}
\usage{
processor(..., prepare, online, offline = NULL)
}
\description{
Wraps all necessary handlers to pipeline function
}
