% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.channels.R
\name{source.channels}
\alias{source.channels}
\title{Wraps raw data stream to use withing pipelinenig functions}
\usage{
source.channels(data, samplingRate, timestamps = NULL)
}
\arguments{
\item{data}{matrix}

\item{timestamps}{You can specify timestamps, otherwise they will be calculated automatically starting from 0}
}
\description{
Wraps raw data stream to use withing pipelinenig functions
}
