% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.events.R
\name{source.events}
\alias{source.events}
\title{Create event stream}
\usage{
source.events(events, timestamps = NULL)
}
\arguments{
\item{events}{List of events (will be coerced to list), must have attribute TS with timestamp or *timestamps* parameter specified}

\item{timestamps}{List of timestamps}
}
\description{
Create event stream
}
