% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/windowize.R
\name{cross.windowizeByEvents}
\alias{cross.windowizeByEvents}
\title{Split data stream to windows using event stream}
\usage{
cross.windowizeByEvents(data, events, windowSize, backBuffer = 10000,
  shift = 0)
}
\arguments{
\item{data}{Data stream}

\item{events}{Events stream}

\item{windowSize}{Size of resulting window}

\item{backBuffer}{Size of buffer for data, may be increased in case of big delay in events arrival}
}
\value{
pipe with type=window
}
\description{
Split data stream to windows using event stream
}

