% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filters.R
\name{pipe.FFTFilter}
\alias{pipe.FFTFilter}
\title{Extracts powers of frequencies in specified channels}
\usage{
pipe.FFTFilter(input, choose)
}
\arguments{
\item{input}{Input pipe}

\item{choose}{DataFrame with columns 'channel' and 'frequency'}
}
\value{
Row where each element corresponds to row from choose and equals to power of channel
}
\description{
Extracts powers of frequencies in specified channels
}

