% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/signalProcessingHelpers.R
\name{signalPreparation}
\alias{signalPreparation}
\title{Typical signal preprocessing for EEG}
\usage{
signalPreparation(input, refs = c(), low = 0.1, high = 40, notch = 50)
}
\description{
Typical signal preprocessing for EEG
}

